+++ src/animation/group.js
@@ -7,8 +7,8 @@ const Util = require('../util/common');
 const Shape = require('../graphic/shape');
 const Timeline = require('../graphic/animate/timeline');
 const Animator = require('../graphic/animate/animator');
-const timeline = Timeline.getGlobalInstance();
 
+let timeline;
 Shape.prototype.animate = function() {
   const attrs = this.get('attrs');
   return new Animator(this, attrs, timeline);
@@ -75,6 +75,10 @@ function getAnimateCfg(geomType, animateCfg) {
 }
 
 module.exports = {
+  afterCanvasInit(/* chart */) {
+    timeline = new Timeline();
+    timeline.play();
+  },
   beforeCanvasDraw(chart) {
     if (chart.get('animate') === false) {
       return;
@@ -104,5 +108,8 @@ module.exports = {
         }
       }
     });
+  },
+  afterCanvasDestroyed(/* chart */) {
+    timeline.stop();
   }
 };