+++ src/graphic/group.js
@@ -24,18 +24,6 @@ class Group extends Element {
     return this;
   }
 
-  // clearTotalMatrix() {
-  //   const m = this._attrs.totalMatrix;
-  //   if (m) {
-  //     this._attrs.totalMatrix = null;
-  //     const children = this._attrs.children;
-  //     for (let i = 0, len = children.length; i < len; i++) {
-  //       const child = children[i];
-  //       child.clearTotalMatrix();
-  //     }
-  //   }
-  // }
-
   /**
    * 获取最小包围盒
    * @return {Object} 返回包围盒