+++ src/plugin/tooltip.js
@@ -1,5 +1,4 @@
 const Util = require('../util/common');
-const DomUtil = require('../util/dom');
 const Global = require('../global');
 const Tooltip = require('../component/tooltip');
 
@@ -174,8 +173,13 @@ class TooltipController {
 
     const chart = self.chart;
     const canvas = chart.get('canvas');
-    const frontPlot = chart.get('frontPlot');
-    const backPlot = chart.get('backPlot');
+    const frontPlot = chart.get('frontPlot').addGroup({
+      className: 'tooltipContainer',
+      zIndex: 10
+    });
+    const backPlot = chart.get('backPlot').addGroup({
+      className: 'tooltipContainer'
+    });
     const plotRange = chart.get('plotRange');
     const coord = chart.get('coord');
 
@@ -378,7 +382,7 @@ class TooltipController {
   handleShowEvent(ev) {
     const chart = this.chart;
     const plot = chart.get('plotRange');
-    const { x, y } = DomUtil.createEvent(ev, chart);
+    const { x, y } = Util.createEvent(ev, chart);
     if (!(x >= plot.tl.x && x <= plot.tr.x && y >= plot.tl.y && y <= plot.br.y)) { // not in chart plot
       this.hideTooltip();
       return;
@@ -408,9 +412,9 @@ class TooltipController {
       if (Util.isFunction(aMethod)) {
         aMethod(method, action); // TODO： 测试，供用户自己绑定事件
       } else if (action === 'bind') {
-        DomUtil.addEventListener(canvasDom, aMethod, method);
+        Util.addEventListener(canvasDom, aMethod, method);
       } else {
-        DomUtil.removeEventListener(canvasDom, aMethod, method);
+        Util.removeEventListener(canvasDom, aMethod, method);
       }
     });
   }
@@ -425,7 +429,7 @@ class TooltipController {
     triggerOff && this._handleEvent(triggerOff, hideMethod, 'bind');
     // TODO: 当用户点击canvas 外的事件时 tooltip 消失
     const docMethod = Util.wrapBehavior(this, 'handleDocEvent');
-    DomUtil.addEventListener(document, 'touchstart', docMethod);
+    document && Util.addEventListener(document, 'touchstart', docMethod);
   }
 
   unBindEvents() {
@@ -438,7 +442,7 @@ class TooltipController {
     triggerOff && this._handleEvent(triggerOff, hideMethod, 'unBind');
     // TODO: 当用户点击canvas 外的事件时 tooltip 消失
     const docMethod = Util.getWrapBehavior(this, 'handleDocEvent');
-    DomUtil.removeEventListener(document, 'touchstart', docMethod);
+    document && Util.removeEventListener(document, 'touchstart', docMethod);
   }
 }
 
@@ -448,10 +452,45 @@ module.exports = {
       chart
     });
     chart.set('tooltipController', tooltipController);
+    /**
+     * 配置 tooltip
+     * @param  {Boolean|Object} enable Boolean 表示是否开启tooltip，Object 则表示配置项
+     * @param  {Object} cfg 配置项
+     * @return {Chart} 返回 Chart 实例
+     */
+    chart.tooltip = function(enable, cfg = {}) {
+      if (Util.isObject(enable)) {
+        cfg = enable;
+        enable = true;
+      }
+      tooltipController.enable = enable;
+      tooltipController.cfg = cfg;
+
+      return this;
+    };
   },
   afterGeomDraw(chart) {
     const tooltipController = chart.get('tooltipController');
     tooltipController.render();
+
+    /**
+     * 根据坐标点显示对应的 tooltip
+     * @param  {Object} point 画布上的点
+     * @return {Chart}       返回 chart 实例
+     */
+    chart.showTooltip = function(point) {
+      tooltipController.showTooltip(point);
+      return this;
+    };
+
+    /**
+     * 隐藏 tooltip
+     * @return {Chart}       返回 chart 实例
+     */
+    chart.hideTooltip = function() {
+      tooltipController.hideTooltip();
+      return this;
+    };
   },
   clearInner(chart) {
     const tooltipController = chart.get('tooltipController');